rename lcd resources
