+++ packages/api/src/index.ts
@@ -1,10 +1,10 @@
-import Service from './service-lcd'
-import Instance from './instance-lcd'
-import Runner from './runner-lcd'
-import Process from './process-lcd'
-import Execution from './execution-lcd'
-import Ownership from './ownership-lcd'
-import Account, { IAccount } from './account-lcd'
+import Service from './service'
+import Instance from './instance'
+import Runner from './runner'
+import Process from './process'
+import Execution from './execution'
+import Ownership from './ownership'
+import Account, { IAccount } from './account'
 import Transaction, { IMsg, IFee } from './transaction'
 import LCDClient from './util/lcd'
